remove mq event listener and add ci web api
